% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OralOpioids.R
\name{Opioid}
\alias{Opioid}
\title{Get the Opioid content from Health Canada by using the DIN or NDC}
\usage{
Opioid(Drug_ID, Opioid_Table)
}
\arguments{
\item{Drug_ID}{A numeric value for the DIN or NDC. Exclude all zeros in front.}

\item{Opioid_Table}{Opioid dataset which can be loaded by using
the \code{load_HealthCanada_Opioid_Table()} or \code{load_FDA_Opioid_Table()} function. The name you use to call the function should be input here.}
}
\value{
Opioid content
}
\description{
\code{Opioid} retrieves the Opioid content from the Opioid_Table.
}
\examples{

US_Opioid_Table <- load_opioid_data("US")
Canada_Opioid_Table <- load_opioid_data("Canada")
Opioid(786535, Canada_Opioid_Table)
Opioid("0093-0058", US_Opioid_Table)
}
