% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OralOpioids.R, R/OralOpioids_ankona - Copy.R,
%   R/OralOpioids_ankona.R
\name{Brand}
\alias{Brand}
\title{Get the Brand name from Health Canada by using the DIN or NDC}
\usage{
Brand(Drug_ID, Opioid_Table)

Brand(Drug_ID, Opioid_Table)

Brand(Drug_ID, Opioid_Table)
}
\arguments{
\item{Drug_ID}{A numeric value for the DIN or NDC. Exclude all zeros in front.}

\item{Opioid_Table}{Opioid dataset which can be loaded by using
the \code{load_HealthCanada_Opioid_Table()} or \code{load_FDA_Opioid_Table()} function. The name you use to call the function should be input here.}
}
\value{
Brand name

Brand name

Brand name
}
\description{
\code{Brand} retrieves the Brand name for an Oral Opioid from the HealthCanada_Opioid_Table.

\code{Brand} retrieves the Brand name for an Oral Opioid from the HealthCanada_Opioid_Table.

\code{Brand} retrieves the Brand name for an Oral Opioid from the HealthCanada_Opioid_Table.
}
\examples{

HealthCanada_Opioid_Table <- load_HealthCanada_Opioid_Table(no_download = TRUE)
Brand(786535, HealthCanada_Opioid_Table)

FDA_Opioid_Table <- load_FDA_Opioid_Table(no_download = TRUE)
Brand("0093-0058", FDA_Opioid_Table)

HealthCanada_Opioid_Table <- load_HealthCanada_Opioid_Table(no_download = TRUE)
Brand(786535, HealthCanada_Opioid_Table)

FDA_Opioid_Table <- load_FDA_Opioid_Table(no_download = TRUE)
Brand("0093-0058", FDA_Opioid_Table)


HealthCanada_Opioid_Table <- load_HealthCanada_Opioid_Table(no_download = TRUE)
Brand(786535, HealthCanada_Opioid_Table)

FDA_Opioid_Table <- load_FDA_Opioid_Table(no_download = TRUE)
Brand("0093-0058", FDA_Opioid_Table)

}
